Merge upload-opendrr-boundaries-sql.yml and upload-geopackages.yml
into new file upload-sqldump-and-geopackages.yml.

The 'prevent-race-condition' concurrency is moved from job level
to workflow level in order to prevent workflow cancellation:

    Canceling since a higher priority waiting request
    for 'prevent-race-condition' exists

Proper workflow_dispatch is also implemented in case there is a need to
manually (re)upload the release assets.

Together with the previous commits, the refactoring of the previous
gpkg-to-pgdump.yml GitHub Actions workflow is now complete.

Fixes #36